extracted packet builder into a dedicated file
